Issue #135 http://code.google.com/p/psutil/issues/detail?id=135
Testing possible fix by changing KERN_ARGMAX sysctl call to use int
instead of size_t for the type. Tested locally, needs to be checked
on OS X 10.7 further to confirm.

+++ psutil/arch/osx/process_info.h
@@ -12,6 +12,6 @@ typedef struct kinfo_proc kinfo_proc;
 
 int get_proc_list(kinfo_proc **procList, size_t *procCount);
 int get_kinfo_proc(pid_t pid, struct kinfo_proc *kp);
-size_t get_argmax(void);
+int get_argmax(void);
 int pid_exists(long pid);
 PyObject* get_arg_list(long pid);
